import-file-text-observables/src/requirements.txt
pycti==3.2.3
PyYAML==5.3.1
urllib3==1.25.9
python-dateutil==2.8.1
beautifulsoup4==4.9.0
requests==2.23.0
stix2==1.4.0
stix==1.2.0.9